slanker/api/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
slowapi==0.1.9
loguru==0.7.2
aiohttp==3.9.1
cryptography==42.0.0
web3==6.15.1
eth-account==0.10.0